BE: clean code & WIP email sending after registration

+++ Spring-Boot/Clinic/backend/clinicalservice/src/main/java/com/nal95/clinic/model/Role.java
@@ -19,8 +19,6 @@ public class Role {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
+    private Long id;
     private String name;
-
 }
